% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBP_CAD.R
\docType{data}
\name{DBP_CAD}
\alias{DBP_CAD}
\title{Genetic association data from diastolic blood pressure (DBP) and coronary artery disease (CAD) GWAS.}
\format{A data frame with 119 rows and 8 variables:
\describe{
  \item{chr.pos}{chromosome position}
  \item{rsid}{RSID}
  \item{bx}{estimated regression coefficient with risk-factor, SBP}
  \item{bxse}{standard error of estimated regression coefficient with risk-factor, SBP}
  \item{by}{estimated regression coefficient with outcome, CAD}
  \item{byse}{standard error of estimated regression coefficient with outcome, CAD}
  \item{a1}{a1 allele}
  \item{a2}{a2 allele}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6284793/}

\url{http://www.phenoscanner.medschl.cam.ac.uk}
}
\usage{
DBP_CAD
}
\description{
A dataset containing chromosome position, rsid and allele information as well as estimates of the regression coefficients and associated standards errors from the SBP and CAD GWAS.
}
\keyword{datasets}
