% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_clust.R
\name{pr_clust}
\alias{pr_clust}
\title{Cluster size and assignemnt probabilities}
\usage{
pr_clust(dta, prob = 0.5, min_obs = 1)
}
\arguments{
\item{dta}{table of results from mr_clust_em$results$best.}

\item{prob}{numeric scalar, keep only variants assigned to clusters above
this allocation probability.}

\item{min_obs}{integer, keep only variants assinged to clusters with more
than or equal to min_obs members.}
}
\value{
The results
}
\description{
Keep results based on a minimum allocation probability and number of
observations in a cluster.
}
